Split UI code to new screen page
